requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==2.2
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
jieba==0.42.1
mysqlclient==2.0.1
Pillow==7.2.0
pycparser==2.20
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.0
python-memcached==1.59
pytz==2020.1
qiniu==7.2.9
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
Whoosh==2.7.4
